requirements.txt
django-tinymce
django
pillow
